more bug fixes re: permutation-corrected fingerprints
this is somewhat of an "aesthetic" correction.  the "old" implementation wasn't *wrong* per se...but i think this version is more intuitive.

in the old implementation, we were effectively taking the *maximum* percentile rank of the observed next recall.

in the new implementation, we're now taking the *average* percentile rank of the observed next recall.

the two are equal, unless there are multiple items with the same feature value.

the most recent commit handles the case where *all* items have the same value (along some feature dimension), but this commit handles the case where *several* (but not necessarily all) items have shared values.  this should give more accurate assessments for categorical features.